// src/generated/Flight.ts
export interface FlightDescriptor {
  type: FlightDescriptor_DescriptorType;
  /**
   * Opaque value used to express a command. Should only be defined when
   * type = CMD.
   */
  cmd: Buffer;
  /**
   * List of strings identifying a particular dataset. Should only be defined
   * when type = PATH.
   */
  path: string[];
}

/** Describes what type of descriptor is defined. */
export enum FlightDescriptor_DescriptorType {
  /** UNKNOWN - Protobuf pattern, not used. */
  UNKNOWN = 0,
  /**
   * PATH - A named path that identifies a dataset. A path is composed of a string
   * or list of strings describing a particular dataset. This is conceptually
   *  similar to a path inside a filesystem.
   */
  PATH = 1,
  /** CMD - An opaque command to generate a dataset. */
  CMD = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface FlightData {
  /**
   * The descriptor of the data. This is only relevant when a client is
   * starting a new DoPut stream.
   */
  flightDescriptor:
    | FlightDescriptor
    | undefined;
  /** Header for message data as described in Message.fbs::Message. */
  dataHeader: Buffer;
  /** Application-defined metadata. */
  appMetadata: Buffer;
  /**
   * The actual batch of Arrow data. Preferably handled with minimal-copies
   * coming last in the definition to help with sidecar patterns (it is
   * expected that some implementations will fetch this field off the wire
   * with specialized code to avoid extra memory copies).
   */
  dataBody: Buffer;
}
// ...
export const FlightDescriptor: MessageFns<FlightDescriptor> = {
  encode(message: FlightDescriptor, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.cmd.length !== 0) {
      writer.uint32(18).bytes(message.cmd);
    }
    for (const v of message.path) {
      writer.uint32(26).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FlightDescriptor {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFlightDescriptor();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.cmd = Buffer.from(reader.bytes());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.path.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FlightDescriptor {
    return {
      type: isSet(object.type) ? flightDescriptor_DescriptorTypeFromJSON(object.type) : 0,
      cmd: isSet(object.cmd) ? Buffer.from(bytesFromBase64(object.cmd)) : Buffer.alloc(0),
      path: globalThis.Array.isArray(object?.path) ? object.path.map((e: any) => globalThis.String(e)) : [],
    };
  },

  toJSON(message: FlightDescriptor): unknown {
    const obj: any = {};
    if (message.type !== 0) {
      obj.type = flightDescriptor_DescriptorTypeToJSON(message.type);
    }
    if (message.cmd.length !== 0) {
      obj.cmd = base64FromBytes(message.cmd);
    }
    if (message.path?.length) {
      obj.path = message.path;
    }
    return obj;
  },

  create(base?: DeepPartial<FlightDescriptor>): FlightDescriptor {
    return FlightDescriptor.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FlightDescriptor>): FlightDescriptor {
    const message = createBaseFlightDescriptor();
    message.type = object.type ?? 0;
    message.cmd = object.cmd ?? Buffer.alloc(0);
    message.path = object.path?.map((e) => e) || [];
    return message;
  },
};
// ...
export const FlightData: MessageFns<FlightData> = {
  encode(message: FlightData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.flightDescriptor !== undefined) {
      FlightDescriptor.encode(message.flightDescriptor, writer.uint32(10).fork()).join();
    }
    if (message.dataHeader.length !== 0) {
      writer.uint32(18).bytes(message.dataHeader);
    }
    if (message.appMetadata.length !== 0) {
      writer.uint32(26).bytes(message.appMetadata);
    }
    if (message.dataBody.length !== 0) {
      writer.uint32(8002).bytes(message.dataBody);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FlightData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFlightData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.flightDescriptor = FlightDescriptor.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.dataHeader = Buffer.from(reader.bytes());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.appMetadata = Buffer.from(reader.bytes());
          continue;
        }
        case 1000: {
          if (tag !== 8002) {
            break;
          }

          message.dataBody = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FlightData {
    return {
      flightDescriptor: isSet(object.flightDescriptor) ? FlightDescriptor.fromJSON(object.flightDescriptor) : undefined,
      dataHeader: isSet(object.dataHeader) ? Buffer.from(bytesFromBase64(object.dataHeader)) : Buffer.alloc(0),
      appMetadata: isSet(object.appMetadata) ? Buffer.from(bytesFromBase64(object.appMetadata)) : Buffer.alloc(0),
      dataBody: isSet(object.dataBody) ? Buffer.from(bytesFromBase64(object.dataBody)) : Buffer.alloc(0),
    };
  },

  toJSON(message: FlightData): unknown {
    const obj: any = {};
    if (message.flightDescriptor !== undefined) {
      obj.flightDescriptor = FlightDescriptor.toJSON(message.flightDescriptor);
    }
    if (message.dataHeader.length !== 0) {
      obj.dataHeader = base64FromBytes(message.dataHeader);
    }
    if (message.appMetadata.length !== 0) {
      obj.appMetadata = base64FromBytes(message.appMetadata);
    }
    if (message.dataBody.length !== 0) {
      obj.dataBody = base64FromBytes(message.dataBody);
    }
    return obj;
  },

  create(base?: DeepPartial<FlightData>): FlightData {
    return FlightData.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FlightData>): FlightData {
    const message = createBaseFlightData();
    message.flightDescriptor = (object.flightDescriptor !== undefined && object.flightDescriptor !== null)
      ? FlightDescriptor.fromPartial(object.flightDescriptor)
      : undefined;
    message.dataHeader = object.dataHeader ?? Buffer.alloc(0);
    message.appMetadata = object.appMetadata ?? Buffer.alloc(0);
    message.dataBody = object.dataBody ?? Buffer.alloc(0);
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}